+++ examples/DoctrineMappedContextTest.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Star\Component\State\Example;
 
@@ -15,6 +15,69 @@ use PHPUnit\Framework\TestCase;
 use Star\Component\State\Builder\StateBuilder;
 use Star\Component\State\StateMetadata;
 
+final class DoctrineMappedContextTest extends TestCase
+{
+    /**
+     * @var EntityManager
+     */
+    private $em;
+
+    public function setUp(): void
+    {
+        if (!\extension_loaded('pdo_sqlite')) {
+            $this->markTestSkipped('Sqlite extension is needed');
+        }
+
+        $configuration = Setup::createAnnotationMetadataConfiguration([__DIR__], true);
+        $this->em = EntityManager::create(
+            [
+                'driver' => 'pdo_sqlite',
+                'in_memory' => true,
+            ],
+            $configuration
+        );
+        $tool = new SchemaTool($this->em);
+        $tool->createSchema([$this->em->getClassMetadata(MyEntity::class)]);
+    }
+
+    public function tearDown(): void
+    {
+        $tool = new SchemaTool($this->em);
+        $tool->dropDatabase();
+    }
+
+    public function test_it_should_keep_the_state_on_hydrate(): void
+    {
+        $beforePersist = new MyEntity();
+
+        $this->assertTrue($beforePersist->isLocked(), 'Should be locked by default');
+        $locked = $this->save($beforePersist);
+        $this->assertTrue($locked->isLocked(), 'Should be locked after persist');
+
+        $locked->unlock();
+
+        $this->assertFalse($locked->isLocked(), 'Should be unlocked before persist');
+        $unlocked = $this->save($locked);
+        $this->assertFalse($unlocked->isLocked(), 'Should be unlocked after persist');
+
+        $unlocked->lock();
+
+        $this->assertTrue($unlocked->isLocked(), 'Should be locked before persist');
+        $final = $this->save($unlocked);
+        $this->assertTrue($final->isLocked(), 'Should be locked after persist');
+    }
+
+    private function save(MyEntity $entity): MyEntity
+    {
+        $this->em->persist($entity);
+        $this->em->flush();
+        $this->em->clear();
+        $this->em->refresh($entity);
+
+        return $entity;
+    }
+}
+
 /**
  * @Entity()
  */
@@ -29,7 +92,7 @@ class MyEntity
     public $id;
 
     /**
-     * @var MyState
+     * @var MyState|StateMetadata
      * @Embedded(class="MyState", columnPrefix="my_")
      */
     private $state;
@@ -39,17 +102,17 @@ class MyEntity
         $this->state = new MyState();
     }
 
-    public function isLocked()
+    public function isLocked(): bool
     {
         return $this->state->isInState('locked');
     }
 
-    public function lock()
+    public function lock(): void
     {
         $this->state = $this->state->transit('lock', $this);
     }
 
-    public function unlock()
+    public function unlock(): void
     {
         $this->state = $this->state->transit('unlock', $this);
     }
@@ -71,73 +134,9 @@ final class MyState extends StateMetadata
         parent::__construct('locked');
     }
 
-    protected function configure(StateBuilder $builder)
+    protected function configure(StateBuilder $builder): void
     {
         $builder->allowTransition('lock', 'unlocked', 'locked');
         $builder->allowTransition('unlock', 'locked', 'unlocked');
-    }
-}
-
-final class DoctrineMappedContextTest extends TestCase
-{
-    /**
-     * @var EntityManager
-     */
-    private $em;
-
-    public function setUp()
-    {
-        $configuration = Setup::createAnnotationMetadataConfiguration([__DIR__], true);
-        $this->em = EntityManager::create(
-            [
-                'driver' => 'pdo_sqlite',
-                'in_memory' => true,
-            ],
-            $configuration
-        );
-        $tool = new SchemaTool($this->em);
-        $tool->createSchema([$this->em->getClassMetadata(MyEntity::class)]);
-    }
-
-    public function tearDown()
-    {
-        $tool = new SchemaTool($this->em);
-        $tool->dropDatabase();
-    }
-
-    public function test_it_should_keep_the_state_on_hydrate()
-    {
-        $beforePersist = new MyEntity();
-
-        $this->assertTrue($beforePersist->isLocked(), 'Should be locked by default');
-        $locked = $this->save($beforePersist);
-        $this->assertTrue($locked->isLocked(), 'Should be locked after persist');
-
-        $locked->unlock();
-
-        $this->assertFalse($locked->isLocked(), 'Should be unlocked before persist');
-        $unlocked = $this->save($locked);
-        $this->assertFalse($unlocked->isLocked(), 'Should be unlocked after persist');
-
-        $unlocked->lock();
-
-        $this->assertTrue($unlocked->isLocked(), 'Should be locked before persist');
-        $final = $this->save($unlocked);
-        $this->assertTrue($final->isLocked(), 'Should be locked after persist');
-    }
-
-    /**
-     * @param MyEntity
-     *
-     * @return MyEntity
-     */
-    private function save($entity)
-    {
-        $this->em->persist($entity);
-        $this->em->flush();
-        $id = $entity->id;
-        $this->em->clear();
-
-        return $this->em->find(MyEntity::class, $id);
     }
 }