+++ src/DuplicateEntryException.php
@@ -1,18 +1,13 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Star\Component\State;
 
 final class DuplicateEntryException extends \LogicException
 {
-    /**
-     * @param string $transition
-     *
-     * @return DuplicateEntryException
-     */
-    public static function duplicateTransition($transition)
+    public static function duplicateTransition(string $transition): self
     {
         return new self(
-            sprintf("The transition '%s' is already registered.", $transition)
+            \sprintf("The transition '%s' is already registered.", $transition)
         );
     }
 }