+++ src/NotFoundException.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /**
  * This file is part of the php-state project.
  *
@@ -9,22 +9,12 @@ namespace Star\Component\State;
 
 final class NotFoundException extends \Exception
 {
-    /**
-     * @param string $name
-     *
-     * @return NotFoundException
-     */
-    public static function stateNotFound($name)
+    public static function stateNotFound(string $name): self
     {
-        return new self(sprintf("The state '%s' could not be found.", $name));
+        return new self(\sprintf("The state '%s' could not be found.", $name));
     }
 
-    /**
-     * @param string $name
-     *
-     * @return NotFoundException
-     */
-    public static function transitionNotFound($name)
+    public static function transitionNotFound(string $name): self
     {
         return new self(sprintf("The transition '%s' could not be found.", $name));
     }