+++ src/StateTransition.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /**
  * This file is part of the php-state project.
  *
@@ -12,15 +12,15 @@ interface StateTransition
     /**
      * @param RegistryBuilder $registry
      */
-    public function onRegister(RegistryBuilder $registry);
+    public function onRegister(RegistryBuilder $registry): void;
 
     /**
      * @return string
      */
-    public function getName();
+    public function getName(): string;
 
     /**
      * @return string
      */
-    public function getDestinationState();
+    public function getDestinationState(): string;
 }