+++ tests/StateMachineTest.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /**
  * This file is part of the php-state project.
  *
@@ -7,6 +7,7 @@
 
 namespace Star\Component\State;
 
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Star\Component\State\Event\StateEventStore;
 use Star\Component\State\Event\TransitionWasFailed;
@@ -32,11 +33,11 @@ final class StateMachineTest extends TestCase
     private $context;
 
     /**
-     * @var EventRegistry|\PHPUnit_Framework_MockObject_MockObject
+     * @var EventRegistry|MockObject
      */
     private $listeners;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->listeners = $this->getMockBuilder(EventRegistry::class)->getMock();
         $this->context = new TestContext('current');
@@ -44,16 +45,14 @@ final class StateMachineTest extends TestCase
         $this->machine = new StateMachine('current', $this->registry, $this->listeners);
     }
 
-    /**
-     * @expectedException        \Star\Component\State\NotFoundException
-     * @expectedExceptionMessage The transition 'not-configured' could not be found.
-     */
-    public function test_it_should_not_allow_to_transition_to_a_not_configured_transition()
+    public function test_it_should_not_allow_to_transition_to_a_not_configured_transition(): void
     {
+        $this->expectException(NotFoundException::class);
+        $this->expectExceptionMessage('The transition \'not-configured\' could not be found.');
         $this->machine->transit('not-configured', $this->context);
     }
 
-    public function test_it_should_transition_from_one_state_to_the_other()
+    public function test_it_should_transition_from_one_state_to_the_other(): void
     {
         $this->registry->addTransition(new OneToOneTransition('name', 'current', 'next'));
         $this->assertTrue($this->machine->isInState('current'));
@@ -63,7 +62,7 @@ final class StateMachineTest extends TestCase
         $this->assertFalse($this->machine->isInState('current'));
     }
 
-    public function test_it_should_trigger_an_event_before_any_transition()
+    public function test_it_should_trigger_an_event_before_any_transition(): void
     {
         $this->listeners
             ->expects($this->at(0))
@@ -77,7 +76,7 @@ final class StateMachineTest extends TestCase
         $this->machine->transit('name', $this->context);
     }
 
-    public function test_it_should_trigger_an_event_after_any_transition()
+    public function test_it_should_trigger_an_event_after_any_transition(): void
     {
         $this->listeners
             ->expects($this->at(1))
@@ -91,38 +90,38 @@ final class StateMachineTest extends TestCase
         $this->machine->transit('name', $this->context);
     }
 
-    /**
-     * @expectedException        \Star\Component\State\InvalidStateTransitionException
-     * @expectedExceptionMessage The transition 't' is not allowed when context 'stdClass' is in state 'current'.
-     */
-    public function test_it_should_throw_exception_with_class_context_when_transition_not_allowed()
+    public function test_it_should_throw_exception_with_class_context_when_transition_not_allowed(): void
     {
         $this->registry->addTransition(new OneToOneTransition('t', 'start', 'end'));
         $this->assertFalse($this->machine->isInState('start'));
 
+        $this->expectException(InvalidStateTransitionException::class);
+        $this->expectExceptionMessage(
+            "The transition 't' is not allowed when context 'stdClass' is in state 'current'."
+        );
         $this->machine->transit('t', new \stdClass);
     }
 
-    /**
-     * @expectedException        \Star\Component\State\InvalidStateTransitionException
-     * @expectedExceptionMessage The transition 'transition' is not allowed when context 'c' is in state 'current'.
-     */
-    public function test_it_should_throw_exception_with_context_as_string_when_transition_not_allowed()
+    public function test_it_should_throw_exception_with_context_as_string_when_transition_not_allowed(): void
     {
         $this->registry->addTransition(new OneToOneTransition('transition', 'start', 'end'));
         $this->assertFalse($this->machine->isInState('start'));
 
+        $this->expectException(InvalidStateTransitionException::class);
+        $this->expectExceptionMessage(
+            "The transition 'transition' is not allowed when context 'c' is in state 'current'."
+        );
         $this->machine->transit('transition', 'c');
     }
 
-    public function test_state_can_have_attribute()
+    public function test_state_can_have_attribute(): void
     {
         $this->registry->registerStartingState('transition', 'current', ['exists']);
         $this->assertFalse($this->machine->hasAttribute('not-exists'));
         $this->assertTrue($this->machine->hasAttribute('exists'));
     }
 
-    public function test_it_should_visit_the_transitions()
+    public function test_it_should_visit_the_transitions(): void
     {
         $registry = $this->getMockBuilder(StateRegistry::class)->getMock();
         $machine = new StateMachine('', $registry, $this->listeners);
@@ -135,16 +134,14 @@ final class StateMachineTest extends TestCase
         $machine->acceptTransitionVisitor($visitor);
     }
 
-    /**
-     * @expectedException        \Star\Component\State\NotFoundException
-     * @expectedExceptionMessage The state 'not-exists' could not be found.
-     */
-    public function test_it_should_throw_exception_when_state_do_not_exists()
+    public function test_it_should_throw_exception_when_state_do_not_exists(): void
     {
+        $this->expectException(NotFoundException::class);
+        $this->expectExceptionMessage("The state 'not-exists' could not be found.");
         $this->machine->isInState('not-exists');
     }
 
-    public function test_it_should_dispatch_an_event_before_a_transition_has_failed()
+    public function test_it_should_dispatch_an_event_before_a_transition_has_failed(): void
     {
         $this->listeners
             ->expects($this->at(1))