+++ tests/Stub/RegistrySpy.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Star\Component\State\Stub;
 
@@ -6,6 +6,9 @@ use Star\Component\State\RegistryBuilder;
 
 final class RegistrySpy implements RegistryBuilder
 {
+    /**
+     * @var string[][][][]
+     */
     public $states = [];
 
     /**
@@ -13,7 +16,7 @@ final class RegistrySpy implements RegistryBuilder
      * @param string $stateName
      * @param string[] $attributes
      */
-    public function registerStartingState($transition, $stateName, array $attributes)
+    public function registerStartingState(string $transition, string $stateName, array $attributes): void
     {
         $this->states[$transition]['start'][][$stateName] = $attributes;
     }
@@ -23,7 +26,7 @@ final class RegistrySpy implements RegistryBuilder
      * @param string $stateName
      * @param string[] $attributes
      */
-    public function registerDestinationState($transition, $stateName, array $attributes)
+    public function registerDestinationState(string $transition, string $stateName, array $attributes): void
     {
         $this->states[$transition]['destination'][][$stateName] = $attributes;
     }